+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.10 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Don't re-use `mapping` variable when migrating portlet data.
+  [witsch]
 
 
 1.9 (2023-05-18)